report/requirements.txt
pandas
python-dotenv
redshift_connector
pytest
pylint